Fix: Store Dto 수정

+++ src/main/java/noodlezip/store/dto/MenuRequestDto.java
@@ -16,8 +16,8 @@ public class MenuRequestDto {
     private String menuDescription;
     private String menuImageUrl;
 
-    private Long ramenCategoryId; // ramenCategory 엔티티의 ID
-    private Long ramenSoupId;     // ramenSoup 엔티티의 ID
+    private Integer ramenCategoryId;
+    private Integer ramenSoupId;
 
     private List<Long> defaultToppingIds; // 기본 토핑 ID 목록 (다대다 관계)
 }
